// src/app/models/empresa.ts
export class Empresa {
    CodigoEmpresa: number;
    CodigoEmpresaTipo: number;
    CodigoEmpresaSituacao: number;
    DtCadastro: Date;
    NomeSocial: string;
    NomeFantasia: string;
    CnpjCpf: string;
    IE: string;
    RgResponsavel: string;
    RgRespOrgaoEmissor: string;
    RgRespOrgaoUFCodigoBaseUF: number;
    RgRespDtEmissao: string;
    CpfRespEmpresa: string;
    NomeResponsavel: string;
    Telefone: string;
    Celular: string;
    Email: string;
    EnderecoRua: string;
    EnderecoNumero: string;
    Bairro: string;
    Cidade: string;
    Cep: string;
    CodigoBaseUF: number;
    CodigoBaseMunicipio: number;
    Municipios: any [];  
    RecordCount: number;
    RowsPage: number;
    PageNumber: number;
}